출석시 Session 모델에 담긴 Code로 출석하도록 수정
- RemoteConfig에 존재하는 공통의 출석코드가 아닌 각 세션에 존재하는 Code를 Input과 비교하여 출석하도록 수정

+++ data/src/main/java/com/yapp/data/datasource/SessionRemoteDataSource.kt
@@ -4,6 +4,6 @@ import com.yapp.data.model.SessionEntity
 
 interface SessionRemoteDataSource {
     suspend fun setSession(session: SessionEntity)
-    suspend fun getSession(id: Long): SessionEntity?
+    suspend fun getSession(id: Int): SessionEntity?
     suspend fun getAllSession(): List<SessionEntity>
 }
